# chore: batch-query charts to then get available filters