@joaoviana 2023-09-25
Still needs: 
- [ ] query batching for explores (or getting common explores)
- [ ] same for other queries

but, for now, this is ok and will be improved by work done on: #7162